+++ Cargo.toml
@@ -3,7 +3,7 @@ resolver = "2"
 members = [
   "sdk/typespec",
   "sdk/typespec/typespec_client_core",
-  "sdk/typespec/typespec_client_core/typespec_client_macros",
+  "sdk/typespec/typespec_derive",
   "sdk/core/azure_core",
   "sdk/core/azure_core_amqp",
   "sdk/identity/azure_identity",
@@ -66,6 +66,7 @@ once_cell = "1.18"
 openssl = { version = "0.10.46" }
 paste = "1.0"
 pin-project = "1.0"
+proc-macro2 = "1.0.86"
 quick-xml = { version = "0.31", features = ["serialize", "serde-types"] }
 rand = "0.8"
 reqwest = { version = "0.12", features = [
@@ -95,6 +96,7 @@ tokio = { version = "1.0", default-features = false, features = [
 tracing = "0.1.40"
 tracing-subscriber = "0.3"
 tz-rs = { version = "0.6" }
+typespec_derive = { path = "sdk/typespec/typespec_derive" }
 url = "2.2"
 uuid = { version = "1.0", features = ["v4"] }
