+++ sdk/identity/azure_identity/Cargo.toml
@@ -13,7 +13,7 @@ categories = ["api-bindings"]
 edition.workspace = true
 
 [dependencies]
-azure_core = { workspace = true }
+azure_core = { workspace = true, features = ["derive"] }
 async-lock = { workspace = true }
 oauth2 = { workspace = true }
 url = { workspace = true }
@@ -25,7 +25,7 @@ async-trait = { workspace = true }
 openssl = { workspace = true, optional = true }
 uuid = { workspace = true, features = ["v4"] }
 pin-project = { workspace = true }
-typespec_client_core = { workspace = true, features = ["derive"] }
+typespec_client_core = { workspace = true }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 async-process = { workspace = true }
