+++ sdk/typespec/typespec_client_core/Cargo.toml
@@ -39,6 +39,8 @@ tokio = { workspace = true, features = ["macros", "rt", "time"] }
 [dev-dependencies]
 once_cell = { workspace = true }
 tokio.workspace = true
+tokio-test.workspace = true
+typespec_derive.workspace = true
 
 [features]
 default = [
