+++ sdk/typespec/typespec_derive/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "typespec_client_macros"
+name = "typespec_derive"
 version = "0.1.0"
 authors.workspace = true
 edition.workspace = true
@@ -16,3 +16,4 @@ proc-macro = true
 [dependencies]
 syn.workspace = true
 quote.workspace = true
+proc-macro2.workspace = true
